requirements.txt
# requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
httpx>=0.25.0